박성철 [Add] 플레이어 특정 각도에서 몸통 회전 구현, idle 상태 구현
