+++ Metalord_btin/MetaLord/Assets/_Test/PSC/Animations/WaitIdle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd0813bbd8d873d46b29714e12e92047
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
