--- Metalord_btin/MetaLord/Assets/_Test/PSC/Scripts/Player/CameraManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85c7a98e48a05d94c9b3abde32bfa3ba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 300
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
