--- Metalord_btin/MetaLord/Assets/_Test/PSC/Scripts/System/Manager/GameSceneManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 795ee2bb46e8cfc4f9c7419536fd187e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
