+++ Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/GunBase.cs
@@ -1,5 +1,4 @@
 using System.Collections;
-using Unity.VisualScripting;
 using UnityEngine;
 
 abstract public class GunBase : MonoBehaviour
@@ -26,7 +25,7 @@ abstract public class GunBase : MonoBehaviour
             brush.splatsY = 4;
         }
     }
-    abstract public void ShootGun();
+    abstract public bool ShootGun();
     
     protected void UsedAmmo(Ray _ray, int _ammo)
     {
