+++ Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/PaintGun.cs
@@ -37,13 +37,13 @@ public class PaintGun : GunBase
     }
 
 
-    public override void ShootGun()
+    public override bool ShootGun()
     {
         if(CheckCanFire() == false || !state.CanFire)
         {
             fireStart = false;
             autotimeCheck = 0f;
-            return;
+            return false;
         }
 
         // 마우스 클릭에서 손을 떼면 사격 중지.
@@ -51,24 +51,30 @@ public class PaintGun : GunBase
         {
             fireStart = false;
             autotimeCheck = 0f;
+            return false;
         }
 
         // 일정시간동안 사격키 입력상태라면 연사모드
         else if (autotimeCheck > state.AutoInitTime && state.CanFire)
         {
             AutoFire();
+            return true;
         }
 
         // 사격을 시작 == 마우스버튼 누른시점동안
         else if (fireStart == true)
         {
             autotimeCheck += Time.deltaTime;
+            return true;
         }
 
         else if (state.reader.ShootKey && state.CanFire)
         {
             NormalFire();
+            return true;
         }
+
+        return false;
     }
