+++ Metalord_btin/MetaLord/ProjectSettings/QualitySettings.asset
@@ -4,7 +4,7 @@
 QualitySettings:
   m_ObjectHideFlags: 0
   serializedVersion: 5
-  m_CurrentQuality: 2
+  m_CurrentQuality: 3
   m_QualitySettings:
   - serializedVersion: 3
     name: Very Low
@@ -107,8 +107,8 @@ QualitySettings:
   - serializedVersion: 3
     name: Medium
     pixelLightCount: 1
-    shadows: 1
-    shadowResolution: 0
+    shadows: 2
+    shadowResolution: 1
     shadowProjection: 1
     shadowCascades: 1
     shadowDistance: 20
@@ -310,7 +310,7 @@ QualitySettings:
     PSP2: 2
     Server: 0
     Stadia: 5
-    Standalone: 2
+    Standalone: 3
     WebGL: 3
     Windows Store Apps: 5
     XboxOne: 5
